refactor(admin-system): move container components to components folder
