+++ modules/admin-system/src/domains/components/domain-detail.component.ts
