+++ modules/admin-system/src/settings/components/setting-detail.component.ts
